disable line overflow highlighting in vim
My previous vim configuration would highlight any line characters
over the 80 character line limit length in an attempt to keep code
files readable.  I've since found that this isn't a realistic limit
for much of the code I open, and it's served as more of a distraction
than a helpful code readability reminder.