increase default font size to improve focus
I have a theory that larger fonts on the terminal will help me focus on the thing
that I'm working on more, since it's going to be that much more in my face.

I am constantly increasing the font size (I should probably get my prescription
checked), so maybe this isn't so bad as a default anyways.