swap super and alt
swapping super with alt, but making the super key the mod key means my
alt key is still my mod

Why do this?  Because firefox and a lot of other apps really love alt
key usage.  But I really love the placement of the alt key on my
keyboard for for sway mod usage.